+++ src/test/java/com/example/noticeboardapi/domain/comment/repository/TreePathCommandRepositoryTest.java
@@ -26,7 +26,7 @@ class TreePathCommandRepositoryTest {
     @Autowired
     private DSLContext dslContext;
 
-    private final static int ROOT_COMMENT_ID = 0;
+    private final static long ROOT_COMMENT_ID = 0;
 
     @Test
     @DisplayName("댓글 트리 저장 테스트")
@@ -35,16 +35,16 @@ class TreePathCommandRepositoryTest {
         Long commentNo = 1001L;
 
         dslContext.insertInto(COMMENT, COMMENT.COMMENT_ID, COMMENT.POST_ID)
-                .values(ULong.valueOf(1001), 1L)
+                .values(1001L, 1L)
                 .execute();
 
         dslContext.insertInto(TREE_PATH, TREE_PATH.POST_ID, TREE_PATH.ANCESTOR, TREE_PATH.DESCENDANT, TREE_PATH.DEPTH)
-                .values(postNo, ULong.valueOf(commentNo), ULong.valueOf(commentNo), 0)
-                .values(postNo, ULong.valueOf(ROOT_COMMENT_ID), ULong.valueOf(commentNo), 1)
+                .values(postNo, commentNo, commentNo, 0)
+                .values(postNo, ROOT_COMMENT_ID, commentNo, 1)
                 .execute();
 
         List<TreePath> treePaths = dslContext.selectFrom(TREE_PATH)
-                .where(TREE_PATH.DESCENDANT.eq(ULong.valueOf(commentNo)))
+                .where(TREE_PATH.DESCENDANT.eq(commentNo))
                 .fetchInto(TreePath.class);
 
         Assertions.assertThat(treePaths.size()).isEqualTo(2);
@@ -58,21 +58,21 @@ class TreePathCommandRepositoryTest {
         Long commentNo = 1001L;
 
         dslContext.insertInto(COMMENT, COMMENT.COMMENT_ID, COMMENT.POST_ID)
-                .values(ULong.valueOf(1001), 1L)
+                .values(1001L, 1L)
                 .execute();
 
         dslContext.insertInto(TREE_PATH)
                 .select(DSL.select(DSL.val(postNo), TREE_PATH.ANCESTOR,
                                 DSL.val(commentNo), TREE_PATH.DEPTH.plus(1))
                         .from(TREE_PATH)
-                        .where(TREE_PATH.DESCENDANT.eq(ULong.valueOf(parentCommentNo)))
-                        .unionAll(DSL.select(DSL.val(postNo), DSL.val(ULong.valueOf(commentNo)),
+                        .where(TREE_PATH.DESCENDANT.eq(parentCommentNo))
+                        .unionAll(DSL.select(DSL.val(postNo), DSL.val(commentNo),
                                 DSL.val(commentNo), DSL.val(0)))
                 ).execute();
 
         //0 -> 894 -> 368 -> 1 -> 1001
         List<TreePath> treePaths = dslContext.selectFrom(TREE_PATH)
-                .where(TREE_PATH.DESCENDANT.eq(ULong.valueOf(commentNo)))
+                .where(TREE_PATH.DESCENDANT.eq(commentNo))
                 .orderBy(TREE_PATH.DEPTH)
                 .fetchInto(TreePath.class);
